pyproject.toml
[project]
name = "meeting-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepgram-sdk==3.10.1",
    "ipykernel>=6.29.5",
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "langchain-openai>=0.3.11",
    "openai>=1.69.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]
